Factor out createAddon for use in heroku-fork
