+++ commands/addons/create.js
@@ -2,7 +2,6 @@
 
 const cli = require('heroku-cli-util')
 const co = require('co')
-let waitForAddonProvisioning = require('../../lib/addons_wait')
 
 function parseConfig (args) {
   let config = {}
@@ -31,58 +30,14 @@ function parseConfig (args) {
   return config
 }
 
-function formatConfigVarsMessage (addon) {
-  let configVars = (addon.config_vars || [])
-
-  if (configVars.length > 0) {
-    configVars = configVars.map(c => cli.color.configVar(c)).join(', ')
-    return `Created ${cli.color.addon(addon.name)} as ${configVars}`
-  } else {
-    return `Created ${cli.color.addon(addon.name)}`
-  }
-}
-
 function * run (context, heroku) {
-  const util = require('../../lib/util')
+  let createAddon = require('../../lib/create_addon')
 
   let {app, flags, args} = context
   let {name, as} = flags
-  let plan = {name: args[0]}
   let config = parseConfig(args.slice(1))
 
-  function createAddon (app, config, name, confirm, plan, as) {
-    return cli.action(`Creating ${plan.name} on ${cli.color.app(app)}`,
-      heroku.post(`/apps/${app}/addons`, {
-        body: { config, name, confirm, plan, attachment: {name: as} },
-        headers: {
-          'accept-expansion': 'plan',
-          'x-heroku-legacy-provider-messages': 'true'
-        }
-      }).then(function (addon) {
-        cli.action.done(cli.color.green(util.formatPrice(addon.plan.price)))
-        return addon
-      })
-    )
-  }
-
-  let addon = yield util.trapConfirmationRequired(context.app, context.flags.confirm, (confirm) => (createAddon(app, config, name, confirm, plan, as)))
-
-  if (addon.provision_message) { cli.log(addon.provision_message) }
-
-  if (addon.state === 'provisioning') {
-    if (context.flags.wait) {
-      cli.log(`Waiting for ${cli.color.addon(addon.name)}...`)
-      addon = yield waitForAddonProvisioning(heroku, addon, 5)
-      cli.log(formatConfigVarsMessage(addon))
-    } else {
-      cli.log(`${cli.color.addon(addon.name)} is being created in the background. The app will restart when complete...`)
-      cli.log(`Use ${cli.color.cmd('heroku addons:info ' + addon.name)} to check creation progress`)
-    }
-  } else if (addon.state === 'deprovisioned') {
-    throw new Error(`The add-on was unable to be created, with status ${addon.state}`)
-  } else {
-    cli.log(formatConfigVarsMessage(addon))
-  }
+  let addon = yield createAddon(heroku, app, args[0], context.flags.confirm, context.flags.wait, {config, name, as})
 
   cli.log(`Use ${cli.color.cmd('heroku addons:docs ' + addon.addon_service.name)} to view documentation`)
 }
