+++ index.js
@@ -24,3 +24,4 @@ exports.commands = flatten([
 ])
 
 exports.resolve = require('./lib/resolve')
+exports.createAddon = require('./lib/create_addon')
